uaa_old/requirements.txt
alembic==1.7.6    
bcrypt==3.2.0    
certifi==2021.10.8
cffi==1.15.0   
charset-normalizer==2.0.11   
click==8.0.3    
Flask==1.1.2    
Flask-Cors==3.0.10   
Flask-Migrate==2.6.0    
flask-paginate==0.8.1    
Flask-PyMongo==2.3.0    
Flask-SQLAlchemy==2.5.1    
greenlet==1.1.2    
idna==3.3      
importlib-metadata==4.10.1   
importlib-resources==5.4.0    
itsdangerous==2.0.1    
Jinja2==3.0.3    
Mako==1.1.6    
MarkupSafe==2.0.1   
psycopg2-binary==2.8.6    
pycparser==2.21     
PyJWT==1.7.1    
pymongo==4.0.1    
requests==2.26.0   
setuptools==44.0.0   
six==1.16.0   
SQLAlchemy==1.4.31   
urllib3==1.26.8   
Werkzeug==1.0.1    
wheel==0.34.2   
zipp==3.7.0    